Work around path issues after Rails 5 upgrade.
I have a strong feeling that there is a better way of doing this.  It seems like
paths for JavaScript and CoffeeScript files within the "Dummy" application's
root are not searched, although I would have expected them to be.  It seems as
if konacha's asset directories are searched instead.  Also, Rails won't load
jQuery using Sprockets for reason.  (Maybe because the application is in a
subdirectory it doesn't know about the jquery-rails gem?)